Modify QueryApp internals, by renaming the role OMC to Data
With this, we do not support the environment variable 'OMC' or the
attribute 'omc' any more, to avoid any confusion.

The proper environment variable to indicate where the input files are
stored is 'DATA', the proper corresponding attribute is 'data'.
The environment variable 'BUREAU' and attribute 'bureau' are still
supported.