Ensure that premium support only versions end up in the proper place
1.0.2 releases aren't going to the public archives since 1.0.2v.
Getting it right required manual intervention since this script wasn't
updated.  Now it is, and we can conduct our 1.0.2 releases anew using
this script.

Doing this required quite a bit of changes, so we maintain a map of
versions <-> series, files <-> versions and series etc, to make sure
the correct directories and announcement mail addresses are affected
for each OpenSSL that's released.

This means that $ftpdir and @distfiles are gone, and are replaced with
the function info(), which returns appropriate information for the
given release series, and %distinfo, which is a comprehensive little
database of collected info that belong together.

^CT-141 Closed