Add github-stats-tools
This is intended as a directory for multiple scripts that pull diverse stats
from github.

The first script is `github-pending.py`, which simply gets the current
amount of open PRs and issues and produces metric samples from that, with
different possible output backends.  This script should be run at regular
intervals, thus producing a timeline of our progress in dealing with PRs and
issues.

It's not yet determined where these data are going to be collected.  There's
a tentative backend to feed a zabbix server, experience will show if that
works as intended or if some other collector would be more suitable.