Modify the CLA check to check CLA: trivial on all commits
The reasoning is that in a PR with several commits, it's possible for
some to be (correctly) marked trivial, and for others not to be (also
correctly).  Does that make the whole PR trivial?  I think not!

Therefore, the whole PR should only be marked trivial if all its commits
are trivial.